requirements.txt
tplinkrouterc6u>=5.18.1
rich>=15.0.0
keyring>=25.0.0
plyer>=2.1.0
qrcode[pil]>=7.4.2
prometheus_client>=0.20.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.0
pystray>=0.19.5
Pillow>=10.0.0
speedtest-cli>=2.1.3
mac-vendor-lookup>=0.1.12
zeroconf>=0.131.0
paho-mqtt>=2.0.0
requests>=2.31.0
cryptography>=42.0.0
# optional — enable rules.yaml + Textual TUI
PyYAML>=6.0
textual>=0.60.0
websockets>=12.0
# dev
pytest>=8.0.0
pyinstaller>=6.0.0
